requirements.txt

svg.py==1.4.2